[C8-17] Improve portability of JVM source code when using gcc7.
This patch places up-to-date type declarations in all places where the gcc switch “-Wno-deprecated-declarations” would flag problems and it enables the switch to catch future related issues. This makes the source code compile on all present Amazon Linux versions.

+++ src/jdk/src/share/native/sun/awt/giflib/dgif_lib.c
@@ -52,9 +52,9 @@ two modules will be linked.  Preserve this property!
 #define UNSIGNED_LITTLE_ENDIAN(lo, hi) ((lo) | ((hi) << 8))
 
 /* avoid extra function call in case we use fread (TVT) */
-#define READ(_gif,_buf,_len)                                     \
-  (((GifFilePrivateType*)_gif->Private)->Read ?                   \
-    ((GifFilePrivateType*)_gif->Private)->Read(_gif,_buf,_len) : \
+#define READ(_gif,_buf,_len)                                             \
+  (((GifFilePrivateType*)_gif->Private)->Read ?                          \
+    (size_t)((GifFilePrivateType*)_gif->Private)->Read(_gif,_buf,_len) : \
     fread(_buf,1,_len,((GifFilePrivateType*)_gif->Private)->File))
 
 static int DGifGetWord(GifFileType *GifFile, GifWord *Word);
@@ -396,7 +396,7 @@ DGifGetImageDesc(GifFileType *GifFile)
     }
     /* Does this image have local color map? */
     if (Buf[0] & 0x80) {
-        unsigned int i;
+        int i;
 
         GifFile->Image.ColorMap = GifMakeMapObject(1 << BitsPerPixel, NULL);
         if (GifFile->Image.ColorMap == NULL) {
